+++ src/mpservice/mpserver.py
@@ -1275,10 +1275,10 @@ class Server:
             book-keeping dict.
 
         .. versionchanged:: 0.13.0
-            To use :meth:`async_call` and :meth:`async_stream`, you must start the object
-            in an **async** context manager. To use :meth:`call` and :meth:`stream`,
-            you must start the object in a (sync) context manager. You can not use
-            both the sync and async methods on one ``Server`` object.
+            The Server object is used in either sync mode or async mode, but not both.
+            Start it in the sync context manager and use it in sync way;
+            start it in the async context manager and use it in async way.
+            In both modes, the main methods are :meth:`call` and :meth:`stream`.
         """
         self.servlet = servlet
 
@@ -1305,6 +1305,8 @@ class Server:
                 cpus = main_cpu
             self._main_cpus = cpus
             psutil.Process().cpu_affinity(cpus=cpus)
+
+        self._async_mode = None
 
     @property
     def capacity(self) -> int:
@@ -1337,8 +1339,7 @@ class Server:
         )
         self._gather_thread.start()
 
-        self.call = self._call
-        self.stream = self._stream
+        self._async_mode = False
 
         return self
 
@@ -1349,13 +1350,58 @@ class Server:
         psutil.Process().cpu_affinity(cpus=[])
         # Reset CPU affinity.
 
+    async def __aenter__(self):
+        self._pipeline_notfull = asyncio.Condition()
+
+        self._q_in = (
+            _SimpleThreadQueue()
+            if self.servlet.input_queue_type == 'thread'
+            else _SimpleProcessQueue()
+        )
+        self._q_out = (
+            _SimpleThreadQueue()
+            if self.servlet.output_queue_type == 'thread'
+            else _SimpleProcessQueue()
+        )
+        self.servlet.start(self._q_in, self._q_out)
+
+        self._gather_task = asyncio.create_task(
+            self._async_gather_output(),
+            name=f"{self.__class__.__name__}._async_gather_output",
+        )
+
+        self._async_mode = True
+
+        return self
+
+    async def __aexit__(self, *args):
+        self._q_in.put(NOMOREDATA)
+        self.servlet.stop()
+        await self._gather_task
+        psutil.Process().cpu_affinity(cpus=[])
+        # Reset CPU affinity.
+
+    def call(self, *args, **kwargs):
+        if self._async_mode is True:
+            return self._async_call(*args, **kwargs)
+        if self._async_mode is False:
+            return self._call(*args, **kwargs)
+        raise RuntimeError("the server is not running")
+
+    def stream(self, *args, **kwargs):
+        if self._async_mode is True:
+            return self._async_stream(*args, **kwargs)
+        if self._async_mode is False:
+            return self._stream(*args, **kwargs)
+        raise RuntimeError("the server is not running")
+
     def _call(self, x, /, *, timeout: int | float = 60, backpressure: bool = True):
         """
         This is called in "embedded" mode for sporadic uses.
         It is not designed to serve high load from multi-thread concurrent
         calls. To process large amounts in embedded model, use :meth:`stream`.
 
-        The parameters ``x`` and ``timeout`` have the same meanings as in :meth:`async_call`.
+        The parameters ``x`` and ``timeout`` have the same meanings as in :meth:`_async_call`.
         """
         fut = self._enqueue(x, timeout, backpressure)
         return self._wait_for_result(fut)
@@ -1465,7 +1511,7 @@ class Server:
             in place of the would-be regular result.
             If ``False``, exceptions will be propagated right away, crashing the program.
         timeout
-            Interpreted the same as in :meth:`call` and :meth:`async_call`.
+            Interpreted the same as in :meth:`_call` and :meth:`_async_call`.
 
             In a streaming task, "timeout" is usually not a concern compared
             to overall throughput. You can usually leave it at the default value.
@@ -1554,38 +1600,6 @@ class Server:
         finally:
             shutdown()
 
-    async def __aenter__(self):
-        self._pipeline_notfull = asyncio.Condition()
-
-        self._q_in = (
-            _SimpleThreadQueue()
-            if self.servlet.input_queue_type == 'thread'
-            else _SimpleProcessQueue()
-        )
-        self._q_out = (
-            _SimpleThreadQueue()
-            if self.servlet.output_queue_type == 'thread'
-            else _SimpleProcessQueue()
-        )
-        self.servlet.start(self._q_in, self._q_out)
-
-        self._gather_task = asyncio.create_task(
-            self._async_gather_output(),
-            name=f"{self.__class__.__name__}._async_gather_output",
-        )
-
-        self.call = self._async_call
-        self.stream = self._async_stream
-
-        return self
-
-    async def __aexit__(self, *args):
-        self._q_in.put(NOMOREDATA)
-        self.servlet.stop()
-        await self._gather_task
-        psutil.Process().cpu_affinity(cpus=[])
-        # Reset CPU affinity.
-
     async def _async_call(
         self, x, /, *, timeout: int | float = 60, backpressure: bool = True
     ):
@@ -1614,7 +1628,7 @@ class Server:
             is over, the :class:`TimeoutError` message will be ".. seconds total".
             This wait, as well as the error message, includes the time that has been spent
             in the "enqueue" step, that is, the timer starts upon receiving the request
-            in ``async_call``.
+            in ``_async_call``.
         backpressure
             If ``True``, and the input queue is full, do not wait; raise :class:`ServerBacklogFull`
             right away. If ``False``, wait on the input queue for as long as